Add support for kotlin and java mixed sources.
This adds a javac step to the KotlincToJarStepFactory.
If any java files are found it will invoke javac, otherwise it won't do
anything.

This also adds java annotations processors to kotlin library rules.
For now, this only works if the java annotation processors do not reference
kotlin code and the kotlin code does not rely on annotation processors.